1. move api route to api.py 2. fix redirect for authenticated user 3. fix flash category
