+++ uwamkp/app.py
@@ -12,6 +12,7 @@ from uwamkp.auth import bp as auth_bp
 from uwamkp.mylisting import bp as mylisting_bp
 from uwamkp.products import bp as products_bp
 from uwamkp.showcase import showcase_bp
+from uwamkp.api import bp as api_bp
 
 
 # create the app and db
@@ -34,6 +35,7 @@ app.register_blueprint(auth_bp)
 app.register_blueprint(mylisting_bp)
 app.register_blueprint(showcase_bp)
 app.register_blueprint(products_bp)
+app.register_blueprint(api_bp)
 
 
 @login_manager.user_loader
