+++ uwamkp/templates/addproduct.html
@@ -56,7 +56,8 @@
                 </div>
             </form>
         </div>
-
+        {% block modal %}
+        {% endblock %}
     </div>
 </div>
