+++ uwamkp/templates/updateproduct.html
@@ -1,7 +1,34 @@
 {% extends "addproduct.html" %}
+
+{% block extra_head %}
+<script src={{ url_for('static', filename='/js/utils.js' ) }}></script>
+<script src={{ url_for('static', filename='/js/updateListing.js' ) }}></script>
+{% endblock %}
+
 {% block submitBtnName %}
     Update
 {% endblock %}
 {% block extraBtn %}
-<button class="btn btn-danger">Delete</button>
+<a class="btn btn-danger" id="btn-deleteListing" deleteUrl={{url_for("api.update_listing", listing_id=listing.id)}}>Delete</a>
 {% endblock %}
+
+{% block modal %}
+<div class="modal fade" id="confirmDelete" tabindex="-1" aria-labelledby="exampleModalLabel" aria-hidden="true">
+    <div class="modal-dialog">
+        <div class="modal-content">
+            <div class="modal-header">
+                <h1 class="modal-title fs-5">Confirm Delete?</h1>
+                <button type="button" class="btn-close" data-bs-dismiss="modal" aria-label="Close"></button>
+            </div>
+            <div class="modal-body">
+                Are you sure you want to delete this Listing?
+            </div>
+            <div class="modal-footer">
+                <button type="button" class="btn btn-secondary" data-bs-dismiss="modal">Close</button>
+                <input id="redirectUrl" value={{url_for("mylisting.my_listing")}} hidden />
+                <button type="button" class="btn btn-danger" id="btn-yes-delete">I'm Sure, Please.</button>
+            </div>
+        </div>
+    </div>
+</div>
+{% endblock%}
